requirements.txt
streamlit
agno
phidata
google-generativeai
Pillow